Loop over a set of coordinates and refresh list of nearby restaurants

+++ WoltLoop/Service/RestaurantsEndpoint.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 enum RestaurantsEndpoint {
-    case nearby(latitude: Float, longitude: Float)
+    case nearby(latitude: Double, longitude: Double)
 }
 
 extension RestaurantsEndpoint: Endpoint {
